Render a react-router Link when adding href to a button
This allows us to render buttons with `<Button href="/annotator" />`,
which will look like a button, but have a link behaviour: users can
middle-click or control click to open the link in a new tab or click the
link to open it in the current page.

We also allow passing target and rel as these can be used for
implementing download links.

Note that ideally we would simply render a `<Link to="/annotator">`
however this would require us to copy over the button styles so that
they work for links as well.
This likely is less maintainable as it could introduce regressions
whenever we update `@adobe/react-spectrum`.