Circumvent security alert by forcing a new version of cookie
The cookie package has a security issue on versions below 0.7.0. This
package is being installed indirectly as a dependency of msw.
Ideally we should update msw to the newest version but this requires
changing a lot of our component and unit tests, which isn't really worth
the change given that we don't use any of msw's cookie functionallity.